Move service2 into fx package
